+++ R/ui_functions.R
@@ -11,7 +11,7 @@ ui_hp_stacked <- function(povline = 1.9,
                           lkup) {
 
   ref_years <- sort(unique(lkup$ref_lkup$reporting_year))
-  ref_years <- ref_years[!ref_years %in% c(1981:1990, 2019)]
+  ref_years <- ref_years[!ref_years %in% c(1981:1989)]
 
   out <- pip_grp(
     country = "all",
@@ -153,19 +153,25 @@ ui_pc_charts <- function(country = c("AGO"),
 #' @inheritParams ui_pc_charts
 #' @return data.table
 #' @export
-ui_pc_regional <- function(country = "all",
-                           year = "all",
-                           povline = 1.9,
+ui_pc_regional <- function(country   = "ALL",
+                           year      = "ALL",
+                           povline   = 1.9,
                            pop_units = 1e6,
                            lkup) {
 
-  out <- pip_grp_logic(country = country,
-                       year    = year,
-                       group_by = "wb",
+  # TEMPORARY UNTIL SELECTION MECHANISM IS BEING IMPROVED
+  country <- toupper(country)
+  if (is.character(year)) {
+    year <- toupper(year)
+  }
+
+  out <- pip_grp_logic(country         = country,
+                       year            = year,
+                       group_by        = "wb",
                        reporting_level = "national",
-                       povline = povline,
-                       lkup = lkup,
-                       censor = TRUE)
+                       povline         = povline,
+                       lkup            = lkup,
+                       censor          = TRUE)
 
   # Add pop_in_poverty and scale according to pop_units
   out$pop_in_poverty <- out$reporting_pop * out$headcount / pop_units
@@ -217,12 +223,16 @@ ui_cp_key_indicators_single <- function(country,
   if (is.null(povline)) {
     poverty_lines <- lkup$pl_lkup$poverty_line
     hc <- lapply(poverty_lines, function(pl) {
-      ui_cp_ki_headcount(country, pl, lkup)
+      ui_cp_ki_headcount(country = country,
+                         povline = pl,
+                         lkup = lkup)
     })
     hc <- data.table::rbindlist(hc)
     # names(hc) <- poverty_lines
   } else {
-    hc <- ui_cp_ki_headcount(country, povline, lkup)
+    hc <- ui_cp_ki_headcount(country = country,
+                             povline = povline,
+                             lkup = lkup)
     # hc <- list(hc)
     # names(hc) <- povline
   }
@@ -243,10 +253,13 @@ ui_cp_key_indicators_single <- function(country,
 #' @inheritParams cp_key_indicators
 #' @return data.table
 #' @noRd
-ui_cp_ki_headcount <- function(country, povline, lkup) {
+ui_cp_ki_headcount <- function(country,
+                               year = "mrv",
+                               povline,
+                               lkup) {
 
   # Fetch most recent year (for CP-display)
-  res <- pip(country = country, year = "mrv",
+  res <- pip(country = country, year = year,
              povline = povline, popshare = NULL,
              welfare_type = "all",
              reporting_level = "all", ppp = NULL,
@@ -548,25 +561,31 @@ ui_svy_meta <- function(country = "all", lkup) {
 #' @return list
 #' @export
 ui_cp_download <- function(country = "AGO",
-                                 povline = NULL,
-                                 lkup) {
+                           year = "ALL",
+                           povline = NULL,
+                           lkup) {
 
   # Select surveys to use for CP page
   lkup$svy_lkup <- lkup$svy_lkup[display_cp == 1]
 
-  if (country == "all") {
+  if (country == "ALL") {
     country_codes <- unique(lkup$svy_lkup$country_code)
     dl <- lapply(country_codes, function(country)
       ui_cp_download_single(
-        country = country, povline = povline, lkup = lkup))
+        country = country,
+        year = year,
+        povline = povline,
+        lkup = lkup))
     dl <- data.table::rbindlist(dl)
     dl <- dl[order(country_code, -reporting_year), ]
   } else {
     dl <- ui_cp_download_single(
-      country = country, povline = povline, lkup = lkup)
+      country = country,
+      year = year,
+      povline = povline,
+      lkup = lkup)
   }
   return(dl)
-
 }
 
 
@@ -579,10 +598,11 @@ ui_cp_download <- function(country = "AGO",
 #' @return list
 #' @keywords internal
 ui_cp_download_single <- function(country,
+                                  year,
                                   povline = 1.9,
                                   lkup) {
 
-  hc <- ui_cp_ki_headcount(country, povline, lkup)
+  hc <- ui_cp_ki_headcount(country, year, povline, lkup)
 
   # Remove "shared_prosperity" since the column names are not consitent with the
   # other data frames
@@ -590,6 +610,7 @@ ui_cp_download_single <- function(country,
   dl <- lapply(indicators, function(x) {
     out <- x[country_code == country, ]
     out[["latest"]] <- NULL
+    out <- unique(out) # HOT FIX: NEED TO FIGURE OUT WHAT THE REAL ISSUE IS
     return(out)
   })